\name{DMarker-package}
\alias{DMarker-package}
\docType{package}
\title{
DMarker
}
\description{
Biomarker discovery is a medical term describing the process by which biomarkers are discovered. Many commonly used blood tests in medicine are biomarkers. There is interest in biomarker discovery on the part of the pharmaceutical industry; blood-test or other biomarkers could serve as intermediate markers of disease in clinical trials, and as possible drug targets. DMarker is a package to predict whether the protein or gene can be detected in blood and urine. \cr
The main function of the package is dmarker(), and the more details you can find in \code{\link{DMarker-method}} \cr
}
\details{
\tabular{ll}{
Package: \tab DMarker\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-05-27\cr
Depends: \tab R (>= 3.0.3), Rcpp (>= 0.11.3)\cr
LinkingTo: \tab Rcpp\cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab yes\cr
LazyData: \tab true\cr
}
}
\author{
Yu Shang (JLU & UGA) \email{yushang@uga.edu} \cr 
Qiong Yu (JLU & UGA) \email{yuqiong@uga.edu} \cr 
Wei Du (JLU & UGA) \email{weidu@uga.edu} \cr 
Ying Xu (JLU & UGA) \email{xyn@bmb.uga.edu} \cr \cr 
Maintainer: Yu Shang (JLU & UGA) \email{yushang@uga.edu} \cr 
}
\references{
\code{[1]} Yan Wang, et al. (2009) \emph{DMarker: A Bio-Marker Inference System for Human Diseases based
on Microarray Gene Expression Data} 2009 \cr
\code{[2]} Juan Cui, et al. (2008) \emph{Computational prediction of human proteins that can be secreted into the bloodstream} BIOINFORMATICS, Vol.24 no. 20 2008 pages 2370-2375 \cr
\code{[3]} Jiaxin Wang, et al. (2013) \emph{Computational Prediction of Human Salivary Proteins from Blood Circulation and Application to Diagnostic Biomarker Identification} PLoS ONE, DOI: 10.1371/journal.pone.0080211, 2013 \cr
\code{[4]} S Hong, et al. (2011) \emph{A Computational Method for Prediction of Excretory Proteins and Application to Identification of cancer markers in urine and application to gastric cancer} PLoS ONE,6(2):e16875, 2011 \cr
\code{[5]} \url{http://bioinfosrv1.bmb.uga.edu/DMarker/} \cr
}
\keyword{ DMarker }
\keyword{ blood }
\keyword{ unine }
\seealso{
\code{\link{DMarker-method}} \code{\link{DMarker}} \code{\link{uniprot}} \cr
}
